RENAME #17 ILogic
