+++ src/logic/ILogic.ts
