+++ src/logic/IMainSceneParams.ts
@@ -1,5 +1,5 @@
-import { ILogic } from "./i-logic";
 import { ICity } from "./ICity";
+import { ILogic } from "./ILogic";
 import { IPlayer } from "./IPlayer";
 
 export interface IMainSceneParams {
